updates face_rec
