+++ butia_recognition/launch/face_rec.launch
@@ -1,0 +1,17 @@
+<launch>
+  <arg name="machine" default="localhost"/>
+  <arg name="use_machine" default="true"/>
+  <arg name="output" default="screen"/>
+
+  <machine name="localhost" address="localhost" if="$(arg use_machine)"/>
+
+  <node pkg="butia_recognition" type="face_recog.py" output="$(arg output)" name="face_recognition_node" machine="$(arg machine)">
+    <rosparam command="load" file="$(find butia_recognition)/config/face_rec.yaml"/>
+  </node>
+  <!-- <node pkg="butia_image2world" type="image2world.py" output="$(arg output)" name="image2world_face_recognition_node" machine="$(arg machine)">
+    <remap from="sub/recognitions2d"  to="/butia_vision/br/face_recognition"/>
+    <remap from="pub/recognitions3d"  to="/butia_vision/br/face_recognition3d"/>
+    <remap from="pub/markers" to="/butia_vision/br/markers"/>
+    <rosparam param="color">[255,0,0]</rosparam>
+ </node> -->
+</launch>
